Modify the entity class variable type to add the entity class database query column name:     Check.java New entity classes:     CheckResult.java     Picture.java     PrimaryCheckSystem.java     ProjectCheckResult.java     SecondaryCheckSystem.java New Mapper class:     PictureMapper.java

+++ src/main/java/com/lot/iotsite/domain/Check.java
@@ -21,7 +21,7 @@ public class Check extends BaseEntity {
     private Long userId;
 
     @TableField("checksys_id")
-    private Integer checkSystemId;
+    private Long checkSystemId;
 
     @TableField("grade")
     private Integer grade;
